fix: Python 3.12 compatibility and LTE RRC OTA v27 support
- Add PY_SSIZE_T_CLEAN macro to all C extension source files
- Cast length arguments to Py_ssize_t in all Py_BuildValue y#/s# calls
- Add LteRrcOtaPacketFmt_v27 format definition and decoding logic
- Fix GUI: populate empty Edit menu with Filter/Search/Time/Reset
- Fix GUI: add data_view guard in OnGridSelect to prevent crashes

+++ dm_collector_c/nr_nas_sm5g_plain_ota_incoming_msg.h
@@ -29,7 +29,7 @@ _decode_nr_nas_sm5g_plain_ota_msg(const char* b, int offset, size_t length,
 
     size_t pdu_length = length - offset;
     PyObject* t = Py_BuildValue("(sy#s)",
-        "Msg", b + offset, pdu_length,
+        "Msg", b + offset, (Py_ssize_t)pdu_length,
         "raw_msg/nas-5gs");
     PyList_Append(result, t);
     Py_DECREF(t);
